backend/requirements.txt
fastapi
uvicorn
pytesseract
pillow
speechrecognition
openai-whisper
torch
sentence-transformers
chromadb
pypdf
gtts
python-dotenv
python-multipart